all baseline pages now feed into the sidebar usign the Baseline object
